4.21 change deletion so only creator of the blog can delete it

+++ part4/blog-list/controllers/blogs.js
@@ -61,9 +61,21 @@ blogRouter.post("/", async (request, response) => {
 });
 
 blogRouter.delete("/:id", async (request, response) => {
-	await Blog.findByIdAndRemove(request.params.id);
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
 
-	response.status(204).end();
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: "token invalid" });
+	}
+	const user = await User.findById(decodedToken.id);
+
+	const blog = await Blog.findById(request.params.id);
+
+	if (blog.user.toString() === user.id.toString()) {
+		await Blog.findByIdAndRemove(request.params.id);
+		response.status(204).end();
+	} else {
+		return response.status(401).json({ error: "unauthorized request" });
+	}
 });
 
 blogRouter.put("/:id", async (request, response) => {
